Fix process_name table
I previously changed this fork of transflac to transcode ogg with ffmpeg instead of oggenc.

I didn't change the corresponding process_name entry.

This meant that transflac didn't sleep when the number of encoding processes is greater than or equal to the number of cpu cores. The result is that it spawns hundreds of processes and hoses the system. This should fix this!