Disable transcoded album art
When transcoding to ogg with ffmpeg a theora album art stream is generated if the source file contains a type 6 (PICTURE) metadata block.

This breaks being able to add replaygain to the ogg file with vorbisgain as it cannot detect the vorbis stream correctly!

Adding '-vn' prevents the processing of video streams and seems to work in my testing.